feat: Support bundling Workflow code prior to Worker creation
