+++ packages/worker/src/index.ts
@@ -8,19 +8,11 @@
  * @module
  */
 
-export {
-  State,
-  Worker,
-  WorkerOptions,
-  CompiledWorkerOptions,
-  ActivityOptions,
-  DataConverter,
-  defaultDataConverter,
-  RetryOptions,
-  errors,
-} from './worker';
+export { State, Worker, ActivityOptions, DataConverter, defaultDataConverter, RetryOptions, errors } from './worker';
+export { WorkerOptions, CompiledWorkerOptions } from './worker-options';
 export { ServerOptions, TLSConfig } from './server-options';
 export { Core } from './core';
 export * from './logger';
 export * from './dependencies';
 export * from './interceptors';
+export { bundleWorkflowCode } from './workflow/bundler';
